Move ActivityStoreContext into android package.
