+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentLifecycleStateTest.kt
@@ -6,7 +6,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
 import com.instacart.formula.fragment.FormulaFragment
 import com.instacart.formula.fragment.FragmentContract
-import com.instacart.formula.integration.ActivityStoreContext
+import com.instacart.formula.android.ActivityStoreContext
 import com.instacart.formula.test.TestContract
 import com.instacart.formula.test.TestContractWithId
 import com.instacart.formula.test.TestFragmentActivity
