+++ formula-android/src/main/java/com/instacart/formula/android/ActivityConfigurator.kt
@@ -2,7 +2,6 @@ package com.instacart.formula.android
 
 import androidx.fragment.app.FragmentActivity
 import com.instacart.formula.integration.ActivityStore
-import com.instacart.formula.integration.ActivityStoreContext
 import kotlin.reflect.KClass
 
 /**
