+++ formula-android/src/main/java/com/instacart/formula/android/ActivityStoreContext.kt
@@ -1,14 +1,14 @@
-package com.instacart.formula.integration
+package com.instacart.formula.android
 
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.Lifecycle
 import com.instacart.formula.activity.ActivityResult
-import com.instacart.formula.android.StreamConfigurator
-import com.instacart.formula.fragment.FragmentContract
 import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.fragment.FragmentFlowStore
 import com.instacart.formula.fragment.FragmentKey
 import com.instacart.formula.fragment.FragmentLifecycleEvent
+import com.instacart.formula.integration.ActivityStore
+import com.instacart.formula.integration.FragmentBindingBuilder
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.disposables.Disposable
