+++ formula-android/src/main/java/com/instacart/formula/android/internal/ActivityStoreContextImpl.kt
@@ -6,7 +6,7 @@ import com.instacart.formula.activity.ActivityResult
 import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.fragment.FragmentKey
 import com.instacart.formula.integration.FragmentId
-import com.instacart.formula.integration.ActivityStoreContext
+import com.instacart.formula.android.ActivityStoreContext
 import com.jakewharton.rxrelay3.BehaviorRelay
 import com.jakewharton.rxrelay3.PublishRelay
 import io.reactivex.rxjava3.core.Observable
